Added context switching interrupt, closes #29

+++ TSOS-2019-master/distrib/os/shell.js
@@ -382,7 +382,7 @@ var TSOS;
                     _CPU.runProcess(pid);
                 }
                 else {
-                    _StdOut.putText("No process with pid: " + pid + " in memory");
+                    _StdOut.putText("No process with pid " + pid + " in memory");
                 }
             }
             else {
@@ -392,6 +392,7 @@ var TSOS;
         shellClearMem(args) {
             _Memory.clearMemory();
             TSOS.Control.updateMemoryDisplay();
+            _StdOut.putText("Memory cleared.");
         }
         shellRunAll(args) {
             _CPU.runAllProcesses();
@@ -418,7 +419,12 @@ var TSOS;
                     _StdOut.putText("pid must be an integer");
                 }
                 else {
-                    _MemoryManager.killProcess(pid);
+                    if (_MemoryManager.doesProcessExist(pid)) {
+                        _MemoryManager.killProcess(pid);
+                    }
+                    else {
+                        _StdOut.putText("Process with pid " + pid + " does not exist.");
+                    }
                 }
             }
         }
@@ -440,6 +446,7 @@ var TSOS;
                 }
                 else {
                     _CpuScheduler.setQuantum(quantum);
+                    _StdOut.putText("Quantum set to " + quantum);
                 }
             }
         }
